pyproject.toml
[tool.poetry]
name = "vividly-dash"
version = "0.1.0"
description = ""
authors = ["[chideracollins]' <[siralexcollins@gmail.com]>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
dash = "^2.18.1"
dash-bootstrap-components = "^1.6.0"
facebook-business = "^21.0.1"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
